Tests were replaced in Class PipelineTest. Class HomeTest was deleted
